refactor: modularization for new architecture
